+++ UnityProject/Assets/Scripts/Core/Addressables/SingletonSOSounds.cs
@@ -88,5 +88,5 @@ public class SingletonSOSounds : SingletonScriptableObject<SingletonSOSounds>
 
    public AddressableAudioSource PosterRipped;
 
-
+   public AddressableAudioSource GlassKnock = null;
 }