+++ UnityProject/Assets/Scripts/Items/Weapons/Gun.cs
@@ -53,6 +53,9 @@ namespace Weapons
 		protected GameObject serverHolder;
 		private RegisterTile shooterRegisterTile;
 
+		protected StandardProgressActionConfig ProgressConfig
+			= new StandardProgressActionConfig(StandardProgressActionType.ItemTransfer);
+
 		/// <summary>
 		/// The current magazine for this weapon, null means empty
 		/// </summary>
@@ -188,11 +191,13 @@ namespace Weapons
 		public ItemSlot pinSlot;
 		public ItemSlot suppressorSlot;
 
+		protected const float PinRemoveTime = 10f;
+
 		// used for clusmy self shooting randomness
 		private System.Random rnd = new System.Random();
 
 		[SyncVar(hook = nameof(SyncPredictionCanFire))]
-		private bool predictionCanFire;
+		protected bool predictionCanFire;
 
 		/// <summary>
 		/// If true, displays a message whenever a gun is shot
@@ -530,13 +535,11 @@ namespace Weapons
 					}
 					else if (Validations.HasItemTrait(interaction.UsedObject, CommonTraits.Instance.Wirecutter) && allowPinSwap)
 					{
-						SyncPredictionCanFire(predictionCanFire, false);
-						Inventory.ServerDrop(pinSlot);
+						PinRemoval(interaction);
 					}
 					else if (Validations.HasItemTrait(interaction.UsedObject, CommonTraits.Instance.FiringPin) && allowPinSwap)
 					{
-						UpdatePredictionCanFire(serverHolder);
-						Inventory.ServerTransfer(interaction.FromSlot, pinSlot);
+						PinAddition(interaction);
 					}
 				}
 				else if (isSuppressed && isSuppressible && suppressorSlot.Item != null)
@@ -631,6 +634,41 @@ namespace Weapons
 					queuedLoadMagNetID = NetId.Invalid;
 				}
 			}
+		}
+
+		protected void PinRemoval(InventoryApply interaction)
+		{
+			void ProgressFinishAction()
+			{
+				Chat.AddActionMsgToChat(interaction.Performer,
+					$"You remove the {FiringPin.gameObject.ExpensiveName()} from {gameObject.ExpensiveName()}",
+					$"{interaction.Performer.ExpensiveName()} removes the {FiringPin.gameObject.ExpensiveName()} from {gameObject.ExpensiveName()}.");
+
+				SyncPredictionCanFire(predictionCanFire, false);
+
+				Inventory.ServerDrop(pinSlot);
+			}
+
+			var bar = StandardProgressAction.Create(ProgressConfig, ProgressFinishAction)
+				.ServerStartProgress(interaction.Performer.RegisterTile(), PinRemoveTime, interaction.Performer);
+
+			if (bar != null)
+			{
+				Chat.AddActionMsgToChat(interaction.Performer,
+					$"You begin removing the {FiringPin.gameObject.ExpensiveName()} from {gameObject.ExpensiveName()}",
+					$"{interaction.Performer.ExpensiveName()} begins removing the {FiringPin.gameObject.ExpensiveName()} from {gameObject.ExpensiveName()}.");
+
+				SoundManager.PlayNetworkedAtPos(SingletonSOSounds.Instance.WireCutter, interaction.Performer.AssumedWorldPosServer(), UnityEngine.Random.Range(0.8f, 1.2f), sourceObj: serverHolder);
+			}
+		}
+
+		protected void PinAddition(InventoryApply interaction)
+		{
+			Chat.AddActionMsgToChat(interaction.Performer,
+				$"You insert the {interaction.UsedObject.gameObject.ExpensiveName()} into {gameObject.ExpensiveName()}.",
+				$"{interaction.Performer.ExpensiveName()} inserts the {interaction.UsedObject.gameObject.ExpensiveName()} into {gameObject.ExpensiveName()}.");
+			UpdatePredictionCanFire(serverHolder);
+			Inventory.ServerTransfer(interaction.FromSlot, pinSlot);
 		}
 
 		/// <summary>
@@ -977,7 +1015,7 @@ namespace Weapons
 		/// <summary>
 		/// Syncs the prediction bool
 		/// </summary>
-		private void SyncPredictionCanFire(bool oldValue, bool newValue)
+		protected void SyncPredictionCanFire(bool oldValue, bool newValue)
 		{
 			predictionCanFire = newValue;
 		}