+++ UnityProject/Assets/Scripts/Tilemaps/Behaviours/TilemapDamage.cs
@@ -78,36 +78,82 @@ public class TilemapDamage : MonoBehaviour, IFireExposable
 			return 0;
 		}
 
-		data.AddTileDamage(Layer.LayerType, basicTile.Armor.GetDamage(damage < basicTile.damageDeflection? 0: damage, attackType));
+		var damageTaken = basicTile.Armor.GetDamage(damage < basicTile.damageDeflection ? 0 : damage, attackType);
 
-		SoundManager.PlayNetworkedAtPos(basicTile.SoundOnHit, worldPosition);
+		data.AddTileDamage(Layer.LayerType, damageTaken);
+
+		if(basicTile.SoundOnHit.AssetAddress != null)
+			SoundManager.PlayNetworkedAtPos(basicTile.SoundOnHit, worldPosition);
+		else{
+			Logger.LogError($"Tried to play SoundOnHit for {basicTile.DisplayName}, but it was null!", Category.Addressables);
+		}
+
+		var totalDamageTaken = data.GetTileDamage(Layer.LayerType);
 
 
-		if (data.GetTileDamage(Layer.LayerType) >= basicTile.MaxHealth)
+		if (totalDamageTaken >= basicTile.MaxHealth)
 		{
 			data.RemoveTileDamage(Layer.LayerType);
 			tileChangeManager.RemoveTile(data.Position, Layer.LayerType);
+
+			//Add new tile if needed
+			//TODO change floors to using overlays, but generic overlay will need to be sprited
+			if (basicTile.ToTileWhenDestroyed != null)
+			{
+				var damageLeft = totalDamageTaken - basicTile.MaxHealth;
+				var tile = basicTile.ToTileWhenDestroyed as BasicTile;
+
+				while (damageLeft > 0 && tile != null)
+				{
+					if (tile.MaxHealth <= damageLeft)
+					{
+						damageLeft -= tile.MaxHealth;
+						tile = tile.ToTileWhenDestroyed as BasicTile;
+					}
+					else
+					{
+						//Atm we just set remaining damage to 0, instead of absorbing it for the new tile
+						damageLeft = 0;
+						tileChangeManager.UpdateTile(data.Position, tile);
+						break;
+					}
+				}
+
+				damageTaken = damageLeft;
+			}
+
+			if (basicTile.SpawnOnDestroy != null)
+			{
+				basicTile.SpawnOnDestroy.SpawnAt(SpawnDestination.At(worldPosition, metaTileMap.ObjectLayer.gameObject.transform));
+			}
+
 			basicTile.LootOnDespawn?.SpawnLoot(worldPosition);
 		}
+		else
+		{
+			if (basicTile.DamageOverlayList != null)
+			{
+				foreach (var overlayData in basicTile.DamageOverlayList.DamageOverlays)
+				{
+					if (overlayData.damagePercentage <= totalDamageTaken / basicTile.MaxHealth)
+					{
+						tileChangeManager.UpdateOverlay(data.Position, overlayData.overlayTile);
+						break;
+					}
+				}
+			}
 
-		return CalculateAbsorbDamaged(attackType,data,basicTile);
-	}
+			//All the damage was absorbed, none left to return for next layer
+			damageTaken = 0;
+		}
 
-	private float CalculateAbsorbDamaged(AttackType attackType, MetaDataNode data, BasicTile basicTile)
-	{
-		var damage = basicTile.MaxHealth - data.GetTileDamage(Layer.LayerType);
-
-		if (basicTile.MaxHealth < damage)
+		if (basicTile.MaxHealth < basicTile.MaxHealth - totalDamageTaken)
 		{
 			data.ResetDamage(Layer.LayerType);
 		}
 
-		if (basicTile.Armor.GetRatingValue(attackType) > 0 && damage > 0)
-		{
-			return damage  / basicTile.Armor.GetRatingValue(attackType);
-		}
-
-		return 0;
+		//Return how much damage is left
+		return damageTaken;
 	}
 
 	public float Integrity(Vector3Int pos)
@@ -121,10 +167,10 @@ public class TilemapDamage : MonoBehaviour, IFireExposable
 		return Mathf.Clamp(layerTile.MaxHealth - metaDataLayer.Get(pos).GetTileDamage(layerTile.LayerType), 0, float.MaxValue);
 	}
 
-	public void RepairWindow(Vector3Int cellPos)
+	public void RemoveTileEffects(Vector3Int cellPos)
 	{
 		var data = metaDataLayer.Get(cellPos);
-		tileChangeManager.RemoveTile(cellPos, LayerType.Effects);
+		tileChangeManager.RemoveOverlay(cellPos, LayerType.Effects);
 		data.ResetDamage(Layer.LayerType);
 	}
 }